Still working on election archiving

+++ app/models/voter_transaction_log.rb
@@ -2,38 +2,25 @@ class VoterTransactionLog < ActiveRecord::Base
   validates_presence_of :origin
   validates_presence_of :datime
   validates_uniqueness_of :origin, :scope => [:datime, :origin_uniq],
-                                   :message => "(+uniq) and daytime duplicated in pre-existing log file."
+                                   :message => "of log file and exact daytime identical to pre-existing log file, duplicate log presumed, rejected."
   belongs_to :election
   has_many :voter_transaction_records, :dependent => :destroy
 
   def delete_archive_file
-    File.delete(self.archive_name)
+    File.delete(self.archive_name) if File.exists?(self.archive_name)
   end
   
   def create_archive_file
-    input_file = self.file_name
     eid = self.election_id
-    e = Election.find(eid)
-    unless e
-      raise Exception, "Election not found during archive operation"
+    unless (e = Election.find(eid))
+      raise Exception, "Election not found during file archive operation"
     end
-    unless (defined?(e.nalllogs) && e.nalllogs.is_a?(Integer))
-      e.nalllogs = e.voter_transaction_logs.length+1 
-    end #JVC invalid in presence of deletes
-    # Either Election needs to hold total number of election logs seen to
-    # date, or it needs to keep the actual list of archived log files, AND, a
-    # log file delete needs to delete the archives file (and maybe the uloads
-    # file?) 
-    path = 'public/archives'
-    file = "e"+eid.to_s+"_vtl_"+e.nalllogs.to_s+".xml"
-    unless File.directory?(path)
+    unless File.directory?(path = 'public/archives')
       raise Exception, "No archive directory: "+path
     end
-    archive_path = path+"/"+file
-    self.archive_name = archive_path
-    upload_path = 'public/uploads/'+self.file_name
-    FileUtils.copy(upload_path,archive_path)
-    e.nalllogs = e.nalllogs+1
+    self.archive_name = path+"/"+"e"+eid.to_s+"_vtl_"+e.logs_max.to_s+".xml"
+    FileUtils.copy('public/uploads/'+self.file_name, self.archive_name)
+    e.log_add(self)
     e.save
   end
